requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
chromadb>=0.4.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
matplotlib>=3.7.0
python-dotenv>=1.0.0